Wrap more libcurl calls
