+++ libmamba/include/mamba/core/context.hpp
@@ -190,7 +190,6 @@ namespace mamba
         bool shell_completion = true;
 
         std::string user_agent = "mamba/" LIBMAMBA_VERSION_STRING;
-        bool curl_initialized = false;
         int connect_timeout_secs = 10;
         // int read_timeout_secs = 60;
         int retry_timeout = 2;  // seconds
