+++ libmamba/include/mamba/core/url.hpp
@@ -66,7 +66,6 @@ namespace mamba
     std::string decode_url(const std::string& url);
     // Only returns a cache name without extension
     std::string cache_name_from_url(const std::string& url);
-    std::string hide_secrets(const std::string_view& str);
 
     class URLHandler
     {
