+++ libmamba/include/mamba/core/util.hpp
@@ -331,6 +331,8 @@ namespace mamba
 
     std::optional<std::string> proxy_match(const std::string& url);
 
+    std::string hide_secrets(const std::string_view& str);
+
     class non_copyable_base
     {
     public:
