+++ libmamba/src/core/fetch.cpp
@@ -30,6 +30,7 @@ namespace mamba
         : m_name(name)
         , m_filename(filename)
         , m_url(unc_url(url))
+        , m_curl_initialized(false)
     {
         m_curl_handle = std::make_unique<CURLHandle>();
         init_curl_ssl();
@@ -47,90 +48,25 @@ namespace mamba
 
     void DownloadTarget::init_curl_handle(CURL* handle, const std::string& url)
     {
-        curl_easy_setopt(handle, CURLOPT_URL, url.c_str());
-        curl_easy_setopt(handle, CURLOPT_NETRC, CURL_NETRC_OPTIONAL);
-        curl_easy_setopt(handle, CURLOPT_FOLLOWLOCATION, 1L);
-
-        // This can improve throughput significantly, see
-        // https://github.com/curl/curl/issues/9601
-        curl_easy_setopt(handle, CURLOPT_BUFFERSIZE, 100 * 1024);
-
-        // DO NOT SET TIMEOUT as it will also take into account multi-start time and
-        // it's just wrong curl_easy_setopt(m_handle, CURLOPT_TIMEOUT,
-        // Context::instance().read_timeout_secs);
-
-        // TODO while libcurl in conda now _has_ http2 support we need to fix mamba to
-        // work properly with it this includes:
-        // - setting the cache stuff correctly
-        // - fixing how the progress bar works
-        curl_easy_setopt(handle, CURLOPT_HTTP_VERSION, CURL_HTTP_VERSION_1_1);
-
         // if the request is slower than 30b/s for 60 seconds, cancel.
         std::string no_low_speed_limit = std::getenv("MAMBA_NO_LOW_SPEED_LIMIT")
                                              ? std::getenv("MAMBA_NO_LOW_SPEED_LIMIT")
                                              : "0";
-        if (no_low_speed_limit == "0")
-        {
-            curl_easy_setopt(handle, CURLOPT_LOW_SPEED_TIME, 60L);
-            curl_easy_setopt(handle, CURLOPT_LOW_SPEED_LIMIT, 30L);
-        }
-
-        curl_easy_setopt(handle, CURLOPT_CONNECTTIMEOUT, Context::instance().connect_timeout_secs);
 
         std::string ssl_no_revoke_env = std::getenv("MAMBA_SSL_NO_REVOKE")
                                             ? std::getenv("MAMBA_SSL_NO_REVOKE")
                                             : "0";
-        if (Context::instance().ssl_no_revoke || ssl_no_revoke_env != "0")
-        {
-            curl_easy_setopt(handle, CURLOPT_SSL_OPTIONS, CURLSSLOPT_NO_REVOKE);
-        }
+        bool set_ssl_no_revoke = (Context::instance().ssl_no_revoke || ssl_no_revoke_env != "0");
 
-        std::optional<std::string> proxy = proxy_match(url);
-        if (proxy)
-        {
-            curl_easy_setopt(handle, CURLOPT_PROXY, proxy->c_str());
-            LOG_INFO << "Using Proxy " << hide_secrets(*proxy);
-        }
-
-        std::string& ssl_verify = Context::instance().ssl_verify;
-        if (ssl_verify.size())
-        {
-            if (ssl_verify == "<false>")
-            {
-                curl_easy_setopt(handle, CURLOPT_SSL_VERIFYPEER, 0L);
-                curl_easy_setopt(handle, CURLOPT_SSL_VERIFYHOST, 0L);
-                if (proxy)
-                {
-                    curl_easy_setopt(handle, CURLOPT_PROXY_SSL_VERIFYPEER, 0L);
-                    curl_easy_setopt(handle, CURLOPT_PROXY_SSL_VERIFYHOST, 0L);
-                }
-            }
-            else if (ssl_verify == "<system>")
-            {
-#ifdef LIBMAMBA_STATIC_DEPS
-                curl_easy_setopt(handle, CURLOPT_CAINFO, nullptr);
-                if (proxy)
-                {
-                    curl_easy_setopt(handle, CURLOPT_PROXY_CAINFO, nullptr);
-                }
-#endif
-            }
-            else
-            {
-                if (!fs::exists(ssl_verify))
-                {
-                    throw std::runtime_error("ssl_verify does not contain a valid file path.");
-                }
-                else
-                {
-                    curl_easy_setopt(handle, CURLOPT_CAINFO, ssl_verify.c_str());
-                    if (proxy)
-                    {
-                        curl_easy_setopt(handle, CURLOPT_PROXY_CAINFO, ssl_verify.c_str());
-                    }
-                }
-            }
-        }
+        curl::set_curl_handle(
+            handle,
+            url,
+            (no_low_speed_limit == "0"),
+            Context::instance().connect_timeout_secs,
+            set_ssl_no_revoke,
+            proxy_match(url),
+            Context::instance().ssl_verify
+        );
     }
 
     int
@@ -159,12 +95,12 @@ namespace mamba
     {
         auto& ctx = Context::instance();
 
-        if (!ctx.curl_initialized)
+        if (!m_curl_initialized)
         {
             if (ctx.ssl_verify == "<false>")
             {
                 LOG_DEBUG << "'ssl_verify' not activated, skipping cURL SSL init";
-                ctx.curl_initialized = true;
+                m_curl_initialized = true;
                 return;
             }
 
@@ -223,7 +159,7 @@ namespace mamba
                 }
             }
 
-            ctx.curl_initialized = true;
+            m_curl_initialized = true;
         }
     }
 
@@ -628,9 +564,9 @@ namespace mamba
             std::stringstream err;
             err << "Download error (" << result << ") " << curl_easy_strerror(result) << " ["
                 << leffective_url << "]\n";
-            if (m_curl_handle->m_errorbuffer[0] != '\0')
+            if (m_curl_handle->get_error_buffer()[0] != '\0')
             {
-                err << m_curl_handle->m_errorbuffer;
+                err << m_curl_handle->get_error_buffer();
             }
             LOG_INFO << err.str();
